do not allow the user to set the file multiple times
The script now checks if the `file` variable has already been set
by a `-f` and exits if it's the case.
`file` has to be set to `""` to avoid an external value to change
the behaviour of the script, e.g. setting `file` as en environment
variable before running the script: behaviour untested.