FEATURE: add flag support to the `gpg-share` script (#48)
* add basic flag support and put the code in main

This commit will NOT work as check_users() needs the users array to
be defined from outside the function and before its definition...
I'll work on that in next commits!

* give real arguments to the functions and remove the exit 1

* pass the users array as argument to functions

This commit also reactivates the script as before.

* give the file and the users with more flags

* show the usage when the script is not used properly

* do not use `echo -n` to get the user when iterating

As the array of users is now properly built and properly passed as an
argument to `check_users` and `encrypt`, there is no need to remove
extra newlines with `echo -n` from inside the loops iterating over the
array of users.

* update the usage string

* update the help message

* fix the error log messages with mentions to the new flags

* require an argument after the `--file` flag

* do not allow the user to set the file multiple times

The script now checks if the `file` variable has already been set
by a `-f` and exits if it's the case.
`file` has to be set to `""` to avoid an external value to change
the behaviour of the script, e.g. setting `file` as en environment
variable before running the script: behaviour untested.

* use `--with-colons` with `gpg` as good practive

A single gpg key has the following form:
```bash
> gpg --list-keys --with-colons user
tru::x:x:x:x:x:x
pub:x:x:x:<small key>:x:x::x:::x:::::x:::x:
fpr:::::::::<pub id>:
uid:x::::x::<uid>::<name>::::::::::x:
sub:x:x:x:<small key>:x:x:::::x:::::x::
fpr:::::::::<sub id>:
```

There is a single "^uid", so we can still count the number of
ids matching the input pattern with a simple `grep "^uid" | wc -l`.

Then, when there is a single user matching the input, we want the
`<pub id>`, i.e. the first "^fpr" after the single "^pub".
And the `<pub id>` appear to be the 10th column.

So the full id-extraction command is
```bash
gpg --quiet --list-keys --with-colons "$user"
| grep "^pub" -A1  # get the pub and the following line
| grep "^fpr"  # get the fpr line
| awk -F ":" '{print $10}'  # get the 10th column
```

The script works just as before.