wrap `$1` inside quotes in the `log_*` functions
This commit helps mitigate the following error
```bash
> DRY_RUN="yes" ./scripts/bump.sh
[i] Current
[-] Exactly
[?] Example:
[?] see
```
i.e. the `log_*` functions only print the first word they are being
given...
This is because of the `log_any ... $1`, as `log_any` takes the `$3` only
and not everything after the fourth argument.

Writing `log_any ... "$1"` makes sure the `$3` argument is made of all the
message and not the first word only!!

With this commit, the output of the same command as above is now
```bash
> DRY_RUN="yes" ./scripts/bump.sh
[i] Current directory is repo's root OK
[-] Exactly one argument is needed, got 0.
[?] Example: `./bump.sh 0.2.6`
[?] see https://github.com/iScsc/iscsc.fr/wiki/Version-bump-procedure#automatic-version-bump for full documentation
```