Migrate database to localhost (#92)
This PR removes the *hardcoded* values of the remote atlas database.
It introduces a new local database deployment in the docker-compose
workflow.
The database is secured thanks to docker networking.

I tested the whole migration :heavy_check_mark: 

:warning: This PR doesn't migrate the current atlas db to the new local
one. This action must be done "by hand" when the website will be bumped.